+++ src/main/java/team05/integrated_feed_backend/module/post/dto/response/PostDto.java
@@ -1,4 +1,4 @@
-package team05.integrated_feed_backend.module.post.dto;
+package team05.integrated_feed_backend.module.post.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
